Finally some POGress with the date parser.

+++ clean-cache.sh
@@ -1,10 +1,16 @@
 #!/bin/env bash
 
-cd src/
+read -p "Deleting __pycache__. Are you sure? [Y/n] " CHOICE
 
-if [ -d __pycache__ ]; then
-	echo "Cleaning __pycache__..."
-	rm -r __pycache__/
+if [ ! $CHOICE == 'n' ]; then
+	cd src/
+
+	if [ -d __pycache__ ]; then
+		echo "Cleaning __pycache__..."
+		rm -r __pycache__/
+	else
+		echo "No cache to clean!"
+	fi
 else
-	echo "No cache to clean!"
+	echo "Deletion aborted."
 fi
